♻️ refactor(analysis): promote TMAnalysisWorkerV2 to TMAnalysisWorker, clean up V1 artifacts
- Move TMAnalysisWorkerV2.php to Analysis/TMAnalysisWorker.php (original namespace)
- Rename class TMAnalysisWorkerV2 → TMAnalysisWorker
- Remove old V1 TMAnalysisWorker.php and 10 V1-only test files
- Add AbstractTest::projectRoot() helper, replace all hardcoded
  dirname/__DIR__ path constructions across 8 test files
- Remove stale PHPStan baseline entry (count 4→0)

2364 tests, 8657 assertions — full green. PHPStan 0 errors.
